+++ src/ggrc/assets/javascripts/components/gdrive/templates/gdrive_picker_launcher_upload_error.mustache
@@ -1,3 +1,8 @@
+{{!
+    Copyright (C) 2017 Google Inc.
+    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+}}
+
 <div>
   <div>Some errors occurred during attaching the files:</div>
   <ul>
