+++ src/ggrc/assets/javascripts/components/gdrive/tests/gdrive_folder_picker_spec.js
